streamlit_vector/requirements.txt
altair==4.2.2
llama-index-core==0.10.67
# see https://llamahub.ai for all integrations
llama-index-embeddings-openai  # can replace with the embed model you want to use
llama-index-llms-openai  # can replace with the llm you want to use
llama-index-readers-file  # needed for loading files
# llama-index-agent-openai  # not needed if not using agents
streamlit==1.19.0
transformers==4.29.2
